# front-end/ng-openapi-gen.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getData","parameters":[],"responses":{"200":{"description":"Returns application data","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Hello API"}}}}}}},"summary":"Get application data","tags":["app"]}},"/api/protected":{"get":{"operationId":"AppController_getProtectedData","parameters":[],"responses":{"200":{"description":"Returns protected data when valid token is provided","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Access granted to protected resource"},"user":{"type":"object","example":{"id":1,"name":"User Name"}},"timestamp":{"type":"string","example":"2025-07-27T10:30:00.000Z"}}}}}},"401":{"description":"Unauthorized - Invalid or missing token"}},"security":[{"access-token":[]}],"summary":"Protected endpoint requiring Bearer token","tags":["app"]}},"/api/users":{"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":"List of all users"}},"summary":"Get all users","tags":["users"]},"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","example":"john_doe"},"email":{"type":"string","example":"john@example.com"},"password":{"type":"string","example":"securePassword123"}},"required":["username","email","password"]}}}},"responses":{"201":{"description":"User created successfully"}},"security":[{"access-token":[]}],"summary":"Create a new user","tags":["users"]}},"/api/users/{id}":{"get":{"operationId":"UserController_getUserById","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User found"},"404":{"description":"User not found"}},"summary":"Get user by ID","tags":["users"]},"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User updated successfully"},"404":{"description":"User not found"}},"security":[{"access-token":[]}],"summary":"Update user","tags":["users"]},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User deleted successfully"},"404":{"description":"User not found"}},"security":[{"access-token":[]}],"summary":"Delete user","tags":["users"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDTO"}}}},"401":{"description":"Invalid credentials"}},"summary":"User login","tags":["auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"201":{"description":"Registration successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenDTO"}}}},"400":{"description":"User already exists"}},"summary":"User registration","tags":["auth"]}},"/api/auth/validate":{"get":{"operationId":"AuthController_validate","parameters":[],"responses":{"200":{"description":""}},"security":[{"access-token":[]}],"summary":"Validate user token","tags":["auth"]}},"/api/pokemon/list":{"get":{"operationId":"PokemonController_getPokemonList","parameters":[{"description":"Name of the Pokémon","required":false,"name":"name","in":"query","schema":{"example":"Bulbasaur","type":"string"}},{"description":"Speed stat","required":false,"name":"speed","in":"query","schema":{"example":45,"type":"number"}},{"description":"Is this Pokémon legendary?","required":false,"name":"legendary","in":"query","schema":{"example":false,"type":"boolean"}},{"description":"Current page number","required":false,"name":"page","in":"query","schema":{"example":1,"type":"number"}},{"description":"Number of items per page","required":false,"name":"limit","in":"query","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":"List of Pokemons","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PokemonListDTO"}}}}},"tags":["Pokemon"]}},"/api/pokemon/import":{"post":{"operationId":"PokemonController_importPokemonCsv","parameters":[],"requestBody":{"required":true,"description":"CSV file to import","content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"204":{"description":"Pokemons imported successfully"}},"tags":["Pokemon"]}},"/api/pokemon-favorite/{pokemonId}":{"post":{"operationId":"PokemonFavoriteController_addFavorite","parameters":[],"requestBody":{"required":true,"description":"DTO for adding a Pokemon to favorites","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddPokemonFavoriteDTO"}}}},"responses":{"201":{"description":"Pokemon added to favorites successfully"},"400":{"description":"Pokemon is already in favorites"}},"tags":["PokemonFavorite"]},"delete":{"operationId":"PokemonFavoriteController_removeFavorite","parameters":[{"name":"pokemonId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Pokemon removed from favorites successfully"},"400":{"description":"Pokemon is not in favorites"}},"tags":["PokemonFavorite"]}},"/api/pokemon-favorite":{"get":{"operationId":"PokemonFavoriteController_getFavorites","parameters":[],"responses":{"200":{"description":""}},"tags":["PokemonFavorite"]}}},"info":{"title":"Pokemon Test API","description":"The Pokemon Test API documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT token","in":"header"}},"schemas":{"LoginDTO":{"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"john_doe"},"password":{"type":"string","description":"Password of the user","example":"securepassword123"}},"required":["username","password"]},"AccessTokenDTO":{"type":"object","properties":{"accessToken":{"type":"string","description":"JWT access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}},"required":["accessToken"]},"RegisterDTO":{"type":"object","properties":{"username":{"type":"string","description":"Username of the user","example":"john_doe"},"password":{"type":"string","description":"Password of the user","example":"securepassword123"}},"required":["username","password"]},"PokemonListDTO":{"type":"object","properties":{}},"AddPokemonFavoriteDTO":{"type":"object","properties":{"pokemonId":{"type":"string","description":"Pokemon ID","example":"uuid"}},"required":["pokemonId"]}}}}